requirements.txt
gradio
pandas
lancedb
uuid
argparse
pathlib
openai
anthropic
groq
instructor
pydantic
pydantic-settings
python-dotenv
kaggle 